adds a unit test using the lock step mechanism
Like that we can control the order of operations being done by the
two threads, and ensure they are both getting the pending objects.
Also we wrap the "handle" method to add a bit of delay, like that we
are sure that the object will not be "completed" before the second
thread is trying to handle it.